+++ api/src/jobs/sync.rs
@@ -5,9 +5,13 @@ use serde::{Deserialize, Serialize};
 use tokio::sync::RwLock;
 use tracing::info;
 
-use universal_inbox::{notification::NotificationSyncSourceKind, user::UserId};
+use universal_inbox::{
+    notification::NotificationSyncSourceKind, task::TaskSyncSourceKind, user::UserId,
+};
 
-use crate::universal_inbox::{notification::service::NotificationService, UniversalInboxError};
+use crate::universal_inbox::{
+    notification::service::NotificationService, task::service::TaskService, UniversalInboxError,
+};
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct SyncNotificationsJob {
@@ -41,6 +45,39 @@ pub async fn handle_sync_notifications(
         service
             .sync_notifications_for_all_users(event.source)
             .await?;
+    }
+
+    Ok(())
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct SyncTasksJob {
+    pub source: Option<TaskSyncSourceKind>,
+    pub user_id: Option<UserId>,
+}
+
+#[tracing::instrument(level = "debug", skip(event, task_service), err)]
+pub async fn handle_sync_tasks(
+    event: SyncTasksJob,
+    task_service: Data<Arc<RwLock<TaskService>>>,
+) -> Result<(), UniversalInboxError> {
+    let source_kind_string = event
+        .source
+        .map(|s| s.to_string())
+        .unwrap_or_else(|| "all types of".to_string());
+    let service = task_service.read().await;
+    if let Some(user_id) = event.user_id {
+        info!("Syncing {source_kind_string} tasks for user {user_id}");
+
+        if let Some(source) = event.source {
+            service.sync_tasks_with_transaction(source, user_id).await?;
+        } else {
+            service.sync_all_tasks(user_id).await?;
+        };
+    } else {
+        info!("Syncing {source_kind_string} tasks for all users");
+
+        service.sync_tasks_for_all_users(event.source).await?;
     }
 
     Ok(())