+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 
 - Use JWT token as access authorization (via a cookie or the `Authorization` header)
 - Introduce ThirdPartyItem entity for Tasks source data
+- Synchronize notifications and tasks on async workers
 
 ### Fixed
 
@@ -52,4 +53,3 @@
   - Link notification to an existing task
 - Act on tasks in the notification list
   - Complete task
-